Changing structure
Taking into account the fact that the Payment Module does not have a complex UI, we choose to implement the front in HTML/CSS/JavaScript,
in order to not allow more complexity than this UI needs. Using React JS for such a simple interface would have been too much. So we 
decided to implement it in those languages, allowing it to be easily integrated to the rest of the app.